feat [PM4-60] show reservations only to users and not to admins
